handle full 64 bit values
add SuInt64
pack/unpack direct to/from int (not via Dnum)

+++ builtin/adler32.go
@@ -62,5 +62,5 @@ func adler32_Value(this Value) Value {
 }
 
 func (h suAdler32) value() Value {
-	return IntVal(int(int32(h.hash.Sum32())))
+	return IntVal(int(h.hash.Sum32()))
 }
